finsh working on rules

+++ src/main/java/com/etz/fraudeagleeyemanager/dto/request/UpdateMapRuleToProductRequest.java
@@ -9,7 +9,7 @@ public class UpdateMapRuleToProductRequest {
 	@NotNull(message="productRuleId cannot be empty")
 	private Long productRuleId;
 	private Boolean notifyAdmin;
-	private Integer emailGroupId;
+	private Long emailGroupId;
 	private Boolean notifyCustomer;
 	@NotNull(message="authorised cannot be empty")
 	private Boolean authorised;
